test auth, add vite proxy for developpement
