+++ frontend/vite.config.ts
@@ -12,5 +12,12 @@ export default defineConfig({
 	],
 	ssr: {
 		noExternal: process.env.NODE_ENV === 'production' ? ['@carbon/charts'] : []
+	},
+	server: {
+		proxy:{
+			'/api': 'http://127.0.0.1:8000',
+			'/auth': 'http://127.0.0.1:8000',
+			'/users': 'http://127.0.0.1:8000',
+		}
 	}
 });
